MR-212 : folder restructuring
